Working on outputting the video to the HTML
I added the entire JS code to the <script> tag because type"module" won't work on my browser without a web server, which isn't something I am interested in using as a dependency because I want to be able to hand this application off to other people without explaining how to download and run one.  Currently, the fix is to change the type to "text/javascript" and this is not pretty and I am sorry.